+++ vitest.config.js
@@ -9,6 +9,7 @@ module.exports = defineConfig({
     setupFiles: ['./src/test/setup.ts'],
     include: ['src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
     globals: true,
+    pool: 'vmThreads',
     coverage: {
       provider: 'v8',
       reporter: ['text', 'json', 'html'],